requirements.txt
streamlit
openai
faiss-cpu
numpy
python-dotenv
pypdf
tqdm
